+++ src/MySQL/ConectWriter.tpp
@@ -9,7 +9,7 @@ namespace ThorsAnvil
     {
 
 template<int len>
-void ConectWriter::writeFixedLengthInteger(unsigned long value)
+void ConectWriter::writeFixedLengthInteger(unsigned long long value)
 {
     THOR_MYSQL_WRITE_INT(value, len);
 }
